sign: move signing funcs to separate files
-signing functions are now residing in sign.h and sign.c
-key functions to do with eckey structure are now in key.c
-added sign.h/.c to Makefile.am and CMakeLists.txt
