+++ CMakeLists.txt
@@ -104,6 +104,7 @@ INSTALL(FILES
     include/dogecoin/random.h
     include/dogecoin/rmd160.h
     include/dogecoin/script.h
+    include/dogecoin/sign.h
     include/dogecoin/serialize.h
     include/dogecoin/sha2.h
     include/dogecoin/tool.h
@@ -151,6 +152,7 @@ TARGET_SOURCES(${LIBDOGECOIN_NAME} PRIVATE
     src/random.c
     src/rmd160.c
     src/script.c
+    src/sign.c
     src/serialize.c
     src/sha2.c
     src/cli/tool.c
