+++ src/address.c
@@ -44,7 +44,6 @@
 #include <dogecoin/bip44.h>
 #include <dogecoin/constants.h>
 #include <dogecoin/chainparams.h>
-#include <dogecoin/ecc.h>
 #include <dogecoin/key.h>
 #include <dogecoin/random.h>
 #include <dogecoin/sha2.h>
@@ -52,198 +51,6 @@
 #include <dogecoin/tool.h>
 #include <dogecoin/utils.h>
 #include <dogecoin/mem.h>
-
-
-/**
- * @brief This function instantiates a new working signature,
- * but does not add it to the hash table.
- * 
- * @return A pointer to the new working signature. 
- */
-signature* new_signature() {
-    signature* sig = (struct signature*)dogecoin_calloc(1, sizeof *sig);
-    sig->recid = 0;
-    sig->idx = HASH_COUNT(signatures) + 1;
-    return sig;
-}
-
-/**
- * @brief This function takes a pointer to an existing working
- * signature object and adds it to the hash table.
- * 
- * @param sig The pointer to the working signature.
- * 
- * @return Nothing.
- */
-void add_signature(signature *sig) {
-    signature* sig_old;
-    HASH_FIND_INT(signatures, &sig->idx, sig_old);
-    if (sig_old == NULL) {
-        HASH_ADD_INT(signatures, idx, sig);
-    } else {
-        HASH_REPLACE_INT(signatures, idx, sig, sig_old);
-    }
-    dogecoin_free(sig_old);
-}
-
-/**
- * @brief This function takes an index and returns the working
- * signature associated with that index in the hash table.
- * 
- * @param idx The index of the target working signature.
- * 
- * @return The pointer to the working signature associated with
- * the provided index.
- */
-signature* find_signature(int idx) {
-    signature* sig;
-    HASH_FIND_INT(signatures, &idx, sig);
-    return sig;
-}
-
-/**
- * @brief This function removes the specified working signature
- * from the hash table and frees the signatures in memory.
- * 
- * @param sig The pointer to the signature to remove.
- * 
- * @return Nothing.
- */
-void remove_signature(signature* sig) {
-    HASH_DEL(signatures, sig); /* delete it (signatures advances to next) */
-    sig->recid = 0;
-    dogecoin_free(sig);
-}
-
-/**
- * @brief This function creates a new sig, places it in
- * the hash table, and returns the index of the new sig,
- * starting from 1 and incrementing each subsequent call.
- * 
- * @return The index of the new sig.
- */
-int start_signature() {
-    signature* sig = new_signature();
-    int index = sig->idx;
-    add_signature(sig);
-    return index;
-}
-
-/**
- * @brief This function frees the memory allocated
- * for an sig.
- * 
- * @param sig The pointer to the sig to be freed.
- * 
- * @return Nothing.
- */
-void free_signature(signature* sig)
-{
-    dogecoin_free(sig->content);
-    dogecoin_free(sig);
-}
-
-/**
- * @brief This function instantiates a new working eckey,
- * but does not add it to the hash table.
- * 
- * @return A pointer to the new working eckey. 
- */
-eckey* new_eckey() {
-    eckey* key = (struct eckey*)dogecoin_calloc(1, sizeof *key);
-    dogecoin_privkey_init(&key->private_key);
-    assert(dogecoin_privkey_is_valid(&key->private_key) == 0);
-    dogecoin_privkey_gen(&key->private_key);
-    assert(dogecoin_privkey_is_valid(&key->private_key)==1);
-    dogecoin_pubkey_init(&key->public_key);
-    dogecoin_pubkey_from_key(&key->private_key, &key->public_key);
-    assert(dogecoin_pubkey_is_valid(&key->public_key) == 1);
-    strcpy(key->public_key_hex, utils_uint8_to_hex((const uint8_t *)&key->public_key, 33));
-    uint8_t pkeybase58c[34];
-    pkeybase58c[0] = dogecoin_chainparams_main.b58prefix_secret_address;
-    pkeybase58c[33] = 1; /* always use compressed keys */
-    memcpy_safe(&pkeybase58c[1], &key->private_key, DOGECOIN_ECKEY_PKEY_LENGTH);
-    assert(dogecoin_base58_encode_check(pkeybase58c, sizeof(pkeybase58c), key->private_key_wif, sizeof(key->private_key_wif)) != 0);
-    key->recid = 0;
-    key->idx = HASH_COUNT(keys) + 1;
-    return key;
-}
-
-/**
- * @brief This function takes a pointer to an existing working
- * eckey object and adds it to the hash table.
- * 
- * @param key The pointer to the working eckey.
- * 
- * @return Nothing.
- */
-void add_eckey(eckey *key) {
-    eckey* key_old;
-    HASH_FIND_INT(keys, &key->idx, key_old);
-    if (key_old == NULL) {
-        HASH_ADD_INT(keys, idx, key);
-    } else {
-        HASH_REPLACE_INT(keys, idx, key, key_old);
-    }
-    dogecoin_free(key_old);
-}
-
-/**
- * @brief This function takes an index and returns the working
- * eckey associated with that index in the hash table.
- * 
- * @param idx The index of the target working eckey.
- * 
- * @return The pointer to the working eckey associated with
- * the provided index.
- */
-eckey* find_eckey(int idx) {
-    eckey* key;
-    HASH_FIND_INT(keys, &idx, key);
-    return key;
-}
-
-/**
- * @brief This function removes the specified working eckey
- * from the hash table and frees the keys in memory.
- * 
- * @param key The pointer to the eckey to remove.
- * 
- * @return Nothing.
- */
-void remove_eckey(eckey* key) {
-    HASH_DEL(keys, key); /* delete it (keys advances to next) */
-    dogecoin_privkey_cleanse(&key->private_key);
-    dogecoin_pubkey_cleanse(&key->public_key);
-    dogecoin_free(key);
-}
-
-/**
- * @brief This function frees the memory allocated
- * for an eckey.
- * 
- * @param eckey The pointer to the eckey to be freed.
- * 
- * @return Nothing.
- */
-void dogecoin_key_free(eckey* eckey)
-{
-    dogecoin_free(eckey);
-}
-
-/**
- * @brief This function creates a new key, places it in
- * the hash table, and returns the index of the new key,
- * starting from 1 and incrementing each subsequent call.
- * 
- * @return The index of the new key.
- */
-int start_key() {
-    eckey* key = new_eckey();
-    int index = key->idx;
-    add_eckey(key);
-    return index;
-}
 
 /**
  * @brief This function generates a new basic public-private
@@ -770,299 +577,4 @@ char* addressFromPrivkey(char* privkey) {
     dogecoin_privkey_cleanse(&key);
     dogecoin_pubkey_cleanse(&pubkey);
     return address;
-}
-
-/**
- * @brief Sign message with private key
- * 
- * @param privkey The key to sign the message with.
- * @param message The message to be signed.
- * 
- * @return Base64 encoded signature if successful, False (0) if not
- * 
- */
-char* signmsgwithprivatekey(char* privkey, char* msg) {
-    if (!privkey || !msg) return false;
-
-    // vars for signing
-    size_t outlen = 74, outlencmp = 64;
-    unsigned char *sig = dogecoin_uchar_vla(outlen), 
-    *sigcmp = dogecoin_uchar_vla(outlencmp);
-
-    // retrieve double sha256 hash of msg:
-    uint256 msgBytes;
-    int ret = dogecoin_dblhash((const unsigned char*)msg, strlen(msg), msgBytes);
-
-    dogecoin_key key;
-    dogecoin_privkey_init(&key);
-    assert(dogecoin_privkey_is_valid(&key) == 0);
-
-    // copy byte array utils_hex_to_uint8(privkey) to dogecoin_key.privkey:
-    memcpy(key.privkey, utils_hex_to_uint8(privkey), strlen(privkey) / 2 + 1);
-    ret = dogecoin_privkey_is_valid(&key);
-    assert(ret == 1);
-
-    // init pubkey
-    dogecoin_pubkey pubkey;
-    dogecoin_pubkey_init(&pubkey);
-    pubkey.compressed = false;
-    ret = dogecoin_pubkey_is_valid(&pubkey);
-    assert(ret == 0);
-    dogecoin_pubkey_from_key(&key, &pubkey);
-    ret = dogecoin_pubkey_is_valid(&pubkey);
-    if (!ret) return 0;
-    ret = dogecoin_privkey_verify_pubkey(&key, &pubkey);
-    if (!ret) return 0;
-
-    // sign hash
-    ret = dogecoin_key_sign_hash(&key, msgBytes, sig, &outlen);
-    if (!ret) {
-        printf("dogecoin_key_sign_hash failed!\n");
-        return 0;
-    }
-    int header = (sig[0] & 0xFF);
-    if (header >= 31) { // this is a compressed key signature
-        pubkey.compressed = true;
-        header -= 24;
-    }
-    int recid = header - 24;
-    // sign compact for recovery of pubkey and free privkey:
-    ret = dogecoin_key_sign_hash_compact_recoverable(&key, msgBytes, sigcmp, &outlencmp, &recid);
-    if (!ret) {
-        printf("key sign recoverable failed!\n");
-        return false;
-    }
-    ret = dogecoin_key_sign_recover_pubkey((const unsigned char*)sigcmp, msgBytes, recid, &pubkey);
-    if (!ret) {
-        printf("key sign recover failed!\n");
-        return false;
-    }
-    ret = dogecoin_pubkey_verify_sig(&pubkey, msgBytes, sig, outlen);
-    if (!ret) {
-        printf("pubkey sig verification failed!\n");
-        return false;
-    }
-
-    dogecoin_free(sigcmp);
-    dogecoin_privkey_cleanse(&key);
-    dogecoin_pubkey_cleanse(&pubkey);
-
-    // base64 encode output and free sig:
-    char* out = b64_encode(sig, outlen);
-    dogecoin_free(sig);
-    return out;
-}
-
-/**
- * @brief Verify signed message using non BIP32 encoded address.
- * 
- * @param address
- * @param sig
- * @param msg
- * 
- * @return P2PKH address if successful, False (0) otherwise
- * 
- */
-char* verifymessage(char* sig, char* msg) {
-    if (!(sig || msg)) return false;
-
-	size_t out_len = b64_decoded_size(sig);
-    unsigned char *out = dogecoin_uchar_vla(out_len), 
-    *sigcomp_out = dogecoin_uchar_vla(65);
-    int ret = b64_decode(sig, out, out_len);
-	if (!ret) {
-        printf("b64_decode failed!\n");
-        return false;
-    }
-
-    // double sha256 hash message:
-    uint256 messageBytes;
-    ret = dogecoin_dblhash((const unsigned char*)msg, strlen(msg), messageBytes);
-    if (!ret) {
-        printf("messageBytes failed\n");
-        return false;
-    }
-    ret = dogecoin_ecc_der_to_compact(out, out_len, sigcomp_out);
-    if (!ret) {
-        printf("ecc der to compact failed!\n");
-        return false;
-    }
-
-    // initialize empty pubkey
-    dogecoin_pubkey pub_key;
-    dogecoin_pubkey_init(&pub_key);
-    pub_key.compressed = false;
-
-    int header = (out[0] & 0xFF);
-
-    if (header >= 31) { // this is a compressed key signature
-        pub_key.compressed = true;
-        header -= 24;
-    }
-
-    int recid = header - 24;
-
-    // recover pubkey
-    ret = dogecoin_key_sign_recover_pubkey((const unsigned char*)sigcomp_out, messageBytes, recid, &pub_key);
-    dogecoin_free(sigcomp_out);
-    if (!ret) {
-        printf("key sign recover failed!\n");
-        return false;
-    }
-    ret = dogecoin_pubkey_verify_sig(&pub_key, messageBytes, out, out_len);
-    dogecoin_free(out);
-    if (!ret) {
-        printf("pubkey sig verification failed!\n");
-        return false;
-    }
-
-    // derive p2pkh address from new injected dogecoin_pubkey with known hexadecimal public key:
-    char* p2pkh_address = dogecoin_char_vla(34 + 1);
-    ret = dogecoin_pubkey_getaddr_p2pkh(&pub_key, &dogecoin_chainparams_main, p2pkh_address);
-    if (!ret) {
-        printf("derived address from pubkey failed!\n");
-        return false;
-    }
-    dogecoin_pubkey_cleanse(&pub_key);
-    return p2pkh_address;
-}
-
-/**
- * @brief Sign message with private key
- * 
- * @param key The key to sign the message with.
- * @param message The message to be signed.
- * 
- * @return Signature struct with base64 encoded signature and recid if successful, False (0) if not
- * 
- */
-signature* signmsgwitheckey(eckey* key, char* msg) {
-    if (!key || !msg) return false;
-
-    // retrieve double sha256 hash of msg:
-    uint256 msgBytes;
-    int ret = dogecoin_dblhash((const unsigned char*)msg, strlen(msg), msgBytes);
-
-    // vars for signing
-    size_t outlen = 74, outlencmp = 64;
-    unsigned char *sig = dogecoin_uchar_vla(outlen), 
-    *sigcmp = dogecoin_uchar_vla(outlencmp);
-
-    // sign hash
-    ret = dogecoin_key_sign_hash(&key->private_key, msgBytes, sig, &outlen);
-    if (!ret) {
-        printf("dogecoin_key_sign_hash failed!\n");
-        return 0;
-    }
-
-    int header = (sig[0] & 0xFF);
-    if (header >= 31) { // this is a compressed key signature
-        header -= 24;
-    }
-
-    int recid = header - 24;
-
-    // sign compact for recovery of pubkey and free privkey:
-    ret = dogecoin_key_sign_hash_compact_recoverable(&key->private_key, msgBytes, sigcmp, &outlencmp, &recid);
-    if (!ret) {
-        printf("key sign recoverable failed!\n");
-        return false;
-    }
-    ret = dogecoin_key_sign_recover_pubkey((const unsigned char*)sigcmp, msgBytes, recid, &key->public_key);
-    if (!ret) {
-        printf("key sign recover failed!\n");
-        return false;
-    }
-    ret = dogecoin_pubkey_verify_sig(&key->public_key, msgBytes, sig, outlen);
-    if (!ret) {
-        printf("pubkey sig verification failed!\n");
-        return false;
-    }
-
-    key->recid = recid;
-    signature* working_sig = new_signature();
-    working_sig->recid = recid;
-
-    // derive p2pkh address from new injected dogecoin_pubkey with known hexadecimal public key:
-    ret = dogecoin_pubkey_getaddr_p2pkh(&key->public_key, &dogecoin_chainparams_main, working_sig->address);
-    if (!ret) {
-        printf("derived address from pubkey failed!\n");
-        return false;
-    }
-    dogecoin_free(sigcmp);
-
-    // base64 encode output and free sig:
-    working_sig->content = b64_encode(sig, outlen);
-
-    dogecoin_free(sig);
-
-    return working_sig;
-}
-
-/**
- * @brief Verify signed message
- * 
- * @param address
- * @param sig
- * @param msg
- * 
- * @return True (1) if successful, False (0) otherwise
- * 
- */
-char* verifymessagewithsig(signature* sig, char* msg) {
-    if (!(sig || msg)) return false;
-
-    char* signature_encoded = sig->content;
-	size_t out_len = b64_decoded_size(signature_encoded);
-    unsigned char *out = dogecoin_uchar_vla(out_len), 
-    *sigcomp_out = dogecoin_uchar_vla(65);
-    int ret = b64_decode(signature_encoded, out, out_len);
-	if (!ret) {
-        printf("b64_decode failed!\n");
-        return false;
-    }
-
-    int recid = sig->recid;
-
-    // double sha256 hash message:
-    uint256 messageBytes;
-    ret = dogecoin_dblhash((const unsigned char*)msg, strlen(msg), messageBytes);
-    if (!ret) {
-        printf("messageBytes failed\n");
-        return false;
-    }
-    ret = dogecoin_ecc_der_to_compact(out, out_len, sigcomp_out);
-    if (!ret) {
-        printf("ecc der to compact failed!\n");
-        return false;
-    }
-
-    dogecoin_pubkey pubkey;
-    dogecoin_pubkey_init(&pubkey);
-    pubkey.compressed = false;
-
-
-    // recover pubkey
-    ret = dogecoin_key_sign_recover_pubkey((const unsigned char*)sigcomp_out, messageBytes, recid, &pubkey);
-    dogecoin_free(sigcomp_out);
-    if (!ret) {
-        printf("key sign recover failed!\n");
-        return false;
-    }
-
-    ret = dogecoin_pubkey_verify_sig(&pubkey, messageBytes, out, out_len);
-    if (!ret) {
-        printf("pubkey sig verification failed!\n");
-        return false;
-    }
-    dogecoin_free(out);
-
-    // derive p2pkh address from new injected dogecoin_pubkey with known hexadecimal public key:
-    char* p2pkh_address = dogecoin_char_vla(34 + 1);
-    ret = dogecoin_pubkey_getaddr_p2pkh(&pubkey, &dogecoin_chainparams_main, p2pkh_address);
-    if (!ret) {
-        printf("derived address from pubkey failed!\n");
-        return false;
-    }
-    return p2pkh_address;
 }
