+++ src/key.c
@@ -44,6 +44,108 @@
 #include <dogecoin/utils.h>
 #include <dogecoin/mem.h>
 
+/**
+ * @brief This function instantiates a new working eckey,
+ * but does not add it to the hash table.
+ * 
+ * @return A pointer to the new working eckey. 
+ */
+eckey* new_eckey() {
+    eckey* key = (struct eckey*)dogecoin_calloc(1, sizeof *key);
+    dogecoin_privkey_init(&key->private_key);
+    assert(dogecoin_privkey_is_valid(&key->private_key) == 0);
+    dogecoin_privkey_gen(&key->private_key);
+    assert(dogecoin_privkey_is_valid(&key->private_key)==1);
+    dogecoin_pubkey_init(&key->public_key);
+    dogecoin_pubkey_from_key(&key->private_key, &key->public_key);
+    assert(dogecoin_pubkey_is_valid(&key->public_key) == 1);
+    strcpy(key->public_key_hex, utils_uint8_to_hex((const uint8_t *)&key->public_key, 33));
+    uint8_t pkeybase58c[34];
+    pkeybase58c[0] = dogecoin_chainparams_main.b58prefix_secret_address;
+    pkeybase58c[33] = 1; /* always use compressed keys */
+    memcpy_safe(&pkeybase58c[1], &key->private_key, DOGECOIN_ECKEY_PKEY_LENGTH);
+    assert(dogecoin_base58_encode_check(pkeybase58c, sizeof(pkeybase58c), key->private_key_wif, sizeof(key->private_key_wif)) != 0);
+    key->recid = 0;
+    key->idx = HASH_COUNT(keys) + 1;
+    return key;
+}
+
+/**
+ * @brief This function takes a pointer to an existing working
+ * eckey object and adds it to the hash table.
+ * 
+ * @param key The pointer to the working eckey.
+ * 
+ * @return Nothing.
+ */
+void add_eckey(eckey *key) {
+    eckey* key_old;
+    HASH_FIND_INT(keys, &key->idx, key_old);
+    if (key_old == NULL) {
+        HASH_ADD_INT(keys, idx, key);
+    } else {
+        HASH_REPLACE_INT(keys, idx, key, key_old);
+    }
+    dogecoin_free(key_old);
+}
+
+/**
+ * @brief This function takes an index and returns the working
+ * eckey associated with that index in the hash table.
+ * 
+ * @param idx The index of the target working eckey.
+ * 
+ * @return The pointer to the working eckey associated with
+ * the provided index.
+ */
+eckey* find_eckey(int idx) {
+    eckey* key;
+    HASH_FIND_INT(keys, &idx, key);
+    return key;
+}
+
+/**
+ * @brief This function removes the specified working eckey
+ * from the hash table and frees the keys in memory.
+ * 
+ * @param key The pointer to the eckey to remove.
+ * 
+ * @return Nothing.
+ */
+void remove_eckey(eckey* key) {
+    HASH_DEL(keys, key); /* delete it (keys advances to next) */
+    dogecoin_privkey_cleanse(&key->private_key);
+    dogecoin_pubkey_cleanse(&key->public_key);
+    dogecoin_free(key);
+}
+
+/**
+ * @brief This function frees the memory allocated
+ * for an eckey.
+ * 
+ * @param eckey The pointer to the eckey to be freed.
+ * 
+ * @return Nothing.
+ */
+void dogecoin_key_free(eckey* eckey)
+{
+    dogecoin_free(eckey);
+}
+
+/**
+ * @brief This function creates a new key, places it in
+ * the hash table, and returns the index of the new key,
+ * starting from 1 and incrementing each subsequent call.
+ * 
+ * @return The index of the new key.
+ */
+int start_key() {
+    eckey* key = new_eckey();
+    int index = key->idx;
+    add_eckey(key);
+    return index;
+}
+
 void dogecoin_privkey_init(dogecoin_key* privkey)
 {
     dogecoin_mem_zero(&privkey->privkey, DOGECOIN_ECKEY_PKEY_LENGTH);
